refactor: move melt_data and clean_year_column functions to common processing module for reuse
