+++ src/data/processing/common_ons_processing.py
@@ -53,3 +53,22 @@ def group_and_sum(df, group_cols=None, sum_col='freq'):
         group_cols = ['ladcode', 'laname', 'year']
     logging.info("Aggregating by: %s", ', '.join(group_cols))
     return df.groupby(group_cols, as_index=False, observed=True).agg({sum_col: 'sum'})
+
+
+def melt_data(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Melt the DataFrame from wide to long format.
+    The 'population_' prefix in the year column is removed.
+    """
+    logging.info("Melting DataFrame from wide to long format...")
+
+    return df.melt(id_vars=["ladcode", "laname", "country", "sex", "age"], var_name="year", value_name="freq")
+
+
+def clean_year_column(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Clean the 'year' column by removing the 'population_' prefix.
+    """
+    logging.info("Cleaning year column...")
+    df['year'] = df['year'].str.replace("population_", "", regex=True)
+    return df
